% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ineacannot.R
\name{ineacannot}
\alias{ineacannot}
\title{Starts the COBECORE meta-data annotator}
\usage{
ineacannot(path = sprintf("\%s/extdata", path.package("ineacannot")))
}
\arguments{
\item{path}{image location path}
}
\description{
The GUI to annotate COBECORE images.
}
\examples{

\dontrun{
# Starts the GUI
ineacannot(path = "~")
}
}
\keyword{GUI,}
\keyword{end,}
\keyword{front}
\keyword{interactive}
