% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top_dir.R
\name{top_dir}
\alias{top_dir}
\title{Find top most directory}
\usage{
top_dir(files)
}
\arguments{
\item{files}{vector with file locations}
}
\description{
From a vector of file locations
}
\keyword{file}
\keyword{io}
\keyword{locations,}
